+++ components/organisms/loginForm/loginForm.scss
@@ -1,0 +1,18 @@
+.login {
+    &__title {
+        margin-left: 1rem;
+    }
+    
+    &__subtitle {
+        margin-left: 1rem;
+    }
+
+    &__form {
+        margin: 1rem;
+        border: 1px solid black;
+
+        &__title {
+            margin-left: 1rem;;
+        }
+    }
+}
