+++ package.json
@@ -17,8 +17,8 @@
     "cp:resetcss": "cp -f node_modules/reboot.css/dist/reboot.min.css utils/resetcss/reboot.min.scss",
     "build": "npm-run-all -p build:*",
     "build:next": "next build",
-    "lint:ts": "eslint {pages,components,utils} --ext .js --ext .jsx --fix",
-    "watch:ts": "chokidar '{components,pages,scss,utils}/**/*.{js,jsx}' -c 'npm run lint:ts'"
+    "lint:js": "eslint {pages,components,utils} --ext .js --ext .jsx --fix",
+    "watch:js": "chokidar '{components,pages,scss,utils}/**/*.{js,jsx}' -c 'npm run lint:js'"
   },
   "dependencies": {
     "google-map-react": "^1.1.5",
